Fold submit relative path back into main submit tests
This was split out to a conditional windows build because I was stumped
on a failing test. Once nywilken fixed it, I realized we could bring it
all back into the same file.
